Add tornado patch, refactor package
Refactored the package so that we have the notion of `init` in the beeline module, in which a libhoney client (added in https://github.com/honeycombio/libhoney-py/pull/33) is initialized with write key, etc. I also followed the ruby beeline model of allowing environment variables to be used in place of hard-coded write key and dataset.

I also wanted to experiment with monkey-patching, since lots of vendors seem to do this. I had previously done some work on tornado, so I used that as an example.

With these changes, the process of instrumenting a tornado webapp would be something like:

```
import beeline
import beeline.patch.tornado

beeline.init(writekey='...', dataset='...')
```

I'd like to try this with other frameworks if it's not too much of a distraction from getting the beeline done.