Add threadlocal event state manager and "add_field" to beeline API
We want the python beeline to be consistent with the other beelines, allowing you to call `beeline.add_field` at arbitrary points in the code to decorate the current event with more fields. Trouble is: how do you know what the current event is? This is where we need some kind of state manager to keep track of the order of events.

The idea is if you are writing instrumentation for the beeline, you just need to call:

```
beeline._new_event({
	"my data": 1,
})
```

and then when wrapping up:

```
beeline._send_event()
```

This will send whatever event was last opened, if there is one. `add_field` will then append the fields to the last opened event. This will work in synchronous frameworks like django and flask, where concurrent requests are handled by identical threads. It won't work for async frameworks, but there are ways to get it done and we can cross that bridge later.