[flask] Add lazy initializer hook for use in uwsgi prefork environments
When Flask users are running in uwsgi with prefork, beeline init has to be handled differently.

If initialization is done at the beginning of the app, the beeline and libhoney client will be initialized, but the worker pool that sends events will be "left behind" in the parent process. The remedy is to initialize the beeline within the worker process. Unfortunately, flask doesn't provide a lot of options to do this. There's `before_first_request` hooks, but this is called *after* the middleware is called, meaning the first event handled by the worker will get dropped.

This adds an optional `init_fn` argument to the middleware, which will be called lazily by the middleware. Users can pass a function that initializes the beeline (or run any code they want, really), which will do a JIT initialization of the beeline before tracing starts.