[tracer] allow the tracer context manager to safely noop if not initialized
As mentioned in https://github.com/honeycombio/beeline-python/issues/51, `tracer` raises
an exception if the beeline has not been initialized. This is the only part of the core API
that will crash an app if `init` has not been called. Add a check for initialization, and return a noop context manager if not initialized.