[traced] defer evaluation of beeline initialization until function is called
While traced was setup to work correctly even if the beeline was uninitialized, the "is initialized" check was evaluated when the wrapped function definition was evaluated rather than when it was called. This meant that when you called `beeline.init` in your program affected whether or not traced functions were actually traced.

This does away with the initialization check when the function is wrapped, instead letting the context manager sort it out at execution time.