Add decorator to pass trace context between threads
Tracing and threading is awkard because of the need to explicitly copy trace state between threadlocal storage. Our docs recommend this approach: https://docs.honeycomb.io/getting-data-in/python/beeline/#threading-and-traces

However, this can be undesirable because it creates a new span at the start of the thread, when the interesting contest is further into the threaded function.

Adding a decorator here to pass trace state over to a threaded function without the added baggage of a new span.