[middleware] add initial Werkzeug support
This is basically a copy of the bottle middleware. It implements the general WSGI middleware interface. Rather than point Werkzeug users to the bottle middleware, we should just have a Werkzeug middleware. I'd like to build on this in the future (adding db instrumentation, etc) but this is a starting point.