Support reinitialization after forking
We have quite a bit of documentation describing how to handle prefork models: https://docs.honeycomb.io/getting-data-in/python/beeline/#using-the-python-beeline-with-python-pre-fork-models

Unfortunately, there are many ways to get it wrong and if you somehow initialize the beeline before the fork and after the fork, your new forked processes will not work correctly because the transmission thread will get left behind with the parent process. Debugging this is painful and we should just work around this scenario by ensuring that calls to `init` are safe no matter where they happen.

To help with this, I am adding code to store the initializing process ID. If a fork happens and the beeline object is copied to the new process via copy-on-write, we'll be able to detect this and correctly initialize the beeline post-fork, including spinning up a new libhoney Client with working transmission thread.

Tested this with a celery app that uses prefork:

```
2019-06-04 17:25:44,503 - honeycomb-sdk - DEBUG - initialized honeycomb client: writekey=changeme dataset=travis.python-testing
2019-06-04 17:25:44,503 - honeycomb-beeline - DEBUG - initialized honeycomb client: writekey=changeme dataset=travis.python-testing service_name=celery
[2019-06-04 17:25:44,717: DEBUG/Worker-2] beeline already initialized, but process ID has changed (was 33775, now 33783). Reinitializing.
[2019-06-04 17:25:44,729: WARNING/Worker-2] 2019-06-04 17:25:44,727 - honeycomb-sdk - DEBUG - initialized honeycomb client: writekey=changeme dataset=travis.python-testing
```