[middleware] Don't try to access self.state.span in handle_error of Flask DB middleware if there is no current_app
Summary: Both `before_cursor_execute` and `after_cursor_execute` are
no-ops if there is no `current_app`, but `handle_error` tries to finish
the span even if there is no `current_app`. This caused an
`AttributeError` when I was using beeline-python in the context of a
multi-threaded application, since when we used `HoneyDBMiddleware` off
of the thread it had been initialized on, `self.state.span` was never
set. We could probably also fix this by using `getattr` on `self.state`
before accessing `span`, but this seems more consistent with how the
other methods operate.