Fix requests patch to correctly build span name
When using the Python `or` keyword in a string interpolation, it behaves
differently to when you use it in assignment. This commit resolves an
instance of this issue in the requests library patch. This helps avoid
situations where the span name would be `requests_None`.

This code snippet demonstrates the issue:

```python

print("Hello %s" % None or "Dave")
# Outputs "Hello None"

name = None or "Dave"
print("Hello %s" % name)
# Outputs "Hello Dave"
```