Fix starlette dependencies
Note that I upgraded to the latest version of poetry (1.1.4), which caused me to have to update some of the other dependencies. Not sure if it was strictly necessary, but best to be on modern versions of everything. Note that I didn’t fix the tests, though - the tests are definitely broken in both 2.7 and 3.8.3