Support streaming responses.
Django streaming responses[1] return their "response," and then
repeatedly yield from their `.streaming_content` iterator to generate
the body.  Ending the request when the response is returned thus omits
all of the time spent streaming the response -- and by then the
HoneyDBWrapper's context has also been exited, which means any
database queries therein are also not traced.

Explicitly examine responses to determine if they are streaming --
and, if so, delay closing the root trace for the response until the
end of the streaming content.

[1] https://docs.djangoproject.com/en/3.2/ref/request-response/#streaminghttpresponse-objects