add rollup SQL duration time to Django traces (#207)
Out of #205 (and internal conversation at my org), adding a rollup field
that includes how much time was spent executing SQL queries was also
desirable for an out-of-the-box experience. This adds that field.

There was a name collision with the already existing `db.duration` here
and the similarly named variable in the Go beeline that is used for
total time, so we choose the `db.total_duration` name here.