add type hints to the traced decorator
We've been adding more type checking our codebase. The beeline.traced
decorator is, unfortunately, hindering that by erasing types on the
functions its applied to. This adds the types it needs.

Along, the way, we need to add `typing_extensions` to provide types
available only in Python 3.10 and later.

The line added is a lil long, but I'm not sure what formatter y'all are
using for this codebase. (Black made a bunch of unrelated changes
happen.)