fix: Only try to calculate query duration when we have start time (#255)
## Which problem is this PR solving?
Occasionally the start time when executing the after cursor handler
doesn't have the start time set. This fix make sure we only attempt to
calculate the duration if we have a start time.

- Fixes #158 

## Short description of the changes
Executing the after cursor handler when a start time is not present
doesn't result in an error.

NOTE: This doesn't attempt to resolve why we sometime execute the after
cursor handler when a start time is not present.